Updating python model
