+++ hdl/rtl/Processor/pkg/CsrDefinitions.vhd
@@ -282,11 +282,11 @@ package body CsrDefinitions is
                 --     end if;
 
                 when 16#323# to 16#33F# =>
-                    hpmaddr := to_natural(i_addr) - 16#B00#;
+                    hpmaddr := to_natural(i_addr) - 16#323#;
                     if (i_wen = '1') then
-                        i_mcsr.mhpmcounters(hpmaddr)(31 downto 0) <= unsigned(i_wdata);
+                        i_mcsr.mhpmevents(hpmaddr)(31 downto 0) <= unsigned(i_wdata);
                     else
-                        o_rdata <= std_logic_vector(i_mcsr.mhpmcounters(hpmaddr)(31 downto 0));
+                        o_rdata <= std_logic_vector(i_mcsr.mhpmevents(hpmaddr)(31 downto 0));
                     end if;
 
                 when 16#340# =>
